pyproject.toml
[project]
name = "eligibility-signposting-api"
version = "0.0.1-alpha"
description = "TODO"
authors = [
  #TODO add authors
]
readme = "README.md"
requires-python = ">=3.13"
dynamic = ["classifiers", "dependencies"]
repository = "https://github.com/NHSDigital/eligibility-signposting-api"
homepage = "https://digital.nhs.uk/developer/api-catalogue"
keywords = ["healthcare", "uk", "nhs", "vaccination", "api"] #TODO add additional keywords

[tool.poetry]
packages = [{include = "eligibility_signposting_api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.13"
flask = {extras = ["async"], version = "^3.1.0"}
httpx = "^0.28.1"
yarl = "^1.18.3"
pydantic = "^2.10.6"
asgiref = "^3.8.1"
boto3 = "^1.37.3"
botocore = "^1.37.3"
eval-type-backport = "^0.2.2"
mangum = "^0.19.0"
wireup = "^1.0.1"
python-json-logger = "^3.3.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.0"
docopt = "^0.6.2"
jsonpath-rw = "^1.4.0"
semver = "^3.0.4"
gitpython = "^3.1.44"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pytest-cov = "^6.0.0"
#pytest-nhsd-apim = "^3.3.2"
aiohttp = "^3.11.12"
awscli = "^1.37.24"
awscli-local = "^0.22.0"
pyhamcrest = "^2.1.0"
factory-boy = "^3.3.3"
pyright = "^1.1.394"
brunns-matchers = "^2.9.0"
localstack = "^4.1.1"
pytest-docker = "^3.2.0"
stamina = "^25.1.0"

[tool.poetry-plugin-lambda-build]
docker-image = "public.ecr.aws/sam/build-python3.13:1.135-x86_64"
docker-network = "host"
docker-platform = "linux/x86_64"
package-artifact-path = "dist/lambda.zip"
without = "dev"

[tool.ruff]
line-length = 120
exclude = ["docs/", "scripts/"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "D"]

[tool.ruff.lint.per-file-ignores]
"src/eligibility_signposting_api/repos/*" = ["ANN401"]
"tests/*" = ["ANN", "INP", "S101", "S106"]

[tool.pyright]
include = ["src/"]
pythonVersion = "3.13"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.run]
relative_files = true
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "raise NotImplementedError",
]
